Move pages into "pages" folder.
