+++ app/(pages)/globals.css
